Updated design of AsyncCommand to handle multiple tasks
Modified IAsyncCommand, AsyncCommand classes to have a RunningTasks property to allow for handling of multiple task execution and allow the property to be used in CanExecute for derived command classes.
Removed ConfigureAwait(false) from implementation code because this really should be done at the library level of an API and not in the application layer.